% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddUnderscoresPipes.R
\name{AddUnderscoresPipes}
\alias{AddUnderscoresPipes}
\title{Put underscores between words in a data frame when separated by pipes.}
\usage{
AddUnderscoresPipes(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A dataframe with underscores between word and spaces after commas.
}
\description{
This function is used to insert underscores between words in a data frame. It
is used primarily to handle columns that respresent responses to survey
questions. This function is specific for columns in which multiple responses
are separated by the pipe character.
}
\examples{
df <- data.frame(full_name = c('John Doe', 'Mary Jane'),
                 activities = c('martial arts | hockey', 'horseback riding'),
                 zip = c(19610, 23302))
AddUnderscoresPipes(df)

}
