% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferencer2.R
\docType{package}
\name{conferencer2}
\alias{conferencer2}
\alias{conferencer2-package}
\title{conferencer2: A package for matching sponsors to the best targets.}
\description{
The conferencr2 package provides functions used to identify the
best matches for conference sponsors based on information contained
in survey data.
}
\section{conferencer2 functions}{

AddUnderscores
AddUnderscoresPipes
GetCompanyMatches
GetMatches
}

