% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyMatchesOutput.R
\name{GetCompanyMatchesOutput}
\alias{GetCompanyMatchesOutput}
\title{Puts the matches for each sponsor into a three-column data frame}
\usage{
GetCompanyMatchesOutput(delegate_data, sponsor_data, Delegates_to_meet)
}
\arguments{
\item{delegate_data}{A data frame.}

\item{sponsor_data}{A data frame. Column 4 contains the name, column 2
contains the email address.}

\item{Delegates_to_meet}{A data frame. Each column corresponds to a sponsor
with the matches listed in rows.}
}
\value{
A three-column dataframe. Each row corresponds to a sponsor.
Matches appear in the third column and are separated by line returns.
}
\description{
This function is used to put the information about matches into a dataframe
with line returns between matches. It is used to produce an output table
that can be easily incorporated into the clients email system.
}
\examples{

}
